Automatically refresh token

+++ BlazorServer/Services/AuthService.cs
@@ -9,12 +9,14 @@ public class AuthService
 {
     private readonly AccessTokenService _accessTokenService;
     private readonly NavigationManager _navigationManager;
+    private readonly RefreshTokenService _refreshTokenService;
     private readonly HttpClient _client;
 
-    public AuthService(AccessTokenService accessTokenService, NavigationManager navigationManager, IHttpClientFactory httpClientFactory)
+    public AuthService(AccessTokenService accessTokenService, NavigationManager navigationManager, IHttpClientFactory httpClientFactory, RefreshTokenService refreshTokenService)
     {
         _accessTokenService = accessTokenService;
         _navigationManager = navigationManager;
+        _refreshTokenService = refreshTokenService;
         _client = httpClientFactory.CreateClient("ApiClient");
     }
 
@@ -22,11 +24,14 @@ public class AuthService
     {
         var status = await _client.PostAsJsonAsync("auth", new { email, password });
 
-        if(status.IsSuccessStatusCode)
+        if (status.IsSuccessStatusCode)
         {
             var token = await status.Content.ReadAsStringAsync();
             var result = JsonSerializer.Deserialize<AuthResponse>(token)!;
+
+            await _accessTokenService.RemoveToken();
             await _accessTokenService.SetToken(result.AccessToken);
+            await _refreshTokenService.SetToken(result.RefreshToken);
             return true;
         }
         else
@@ -35,9 +40,36 @@ public class AuthService
         }
     }
 
+    public async Task<bool> RefreshTokenAsync()
+    {
+        var refreshToken = await _refreshTokenService.GetToken();
+        _client.DefaultRequestHeaders.Add("Cookie", $"refreshToken={refreshToken}");
+        var responseMessage = await _client.PostAsync("auth/refresh", null);
+        if (responseMessage.IsSuccessStatusCode)
+        {
+            var token = await responseMessage.Content.ReadAsStringAsync();
+
+            if (!string.IsNullOrEmpty(token))
+            {
+                var result = JsonSerializer.Deserialize<AuthResponse>(token)!;
+                await _accessTokenService.SetToken(result.AccessToken);
+                await _refreshTokenService.SetToken(result.RefreshToken);
+                return true;
+            }
+        }
+
+        return false;
+    }
     public async Task Logout()
     {
-        await _accessTokenService.RemoveToken();
-        _navigationManager.NavigateTo("/login");
+        var refreshToken = await _refreshTokenService.GetToken();
+        _client.DefaultRequestHeaders.Add("Cookie", $"refreshToken={refreshToken}");
+        var responseMessage = await _client.PostAsync("auth/logout", null);
+        if (responseMessage.IsSuccessStatusCode)
+        {
+            await _accessTokenService.RemoveToken();
+            await _refreshTokenService.Remove();
+            _navigationManager.NavigateTo("/login", forceLoad: true);
+        }
     }
 }
